Use [`ValueListBuilder`](https://github.com/dotnet/runtime/blob/7060c62203d056f074e7d43f96fd43c37a19d4a2/src/libraries/System.Private.CoreLib/src/System/Collections/Generic/ValueListBuilder.cs) to remove `List<Doc>` allocations. 

- I'm getting around the limit on `stackalloc` for ref types by letting the collection expression for `Span<Doc>` handle the unsafe code to initalise and convert a struct into a `Span`.
- Opted to go with 8 items as my inital size as my testing showed that it never exceeds this value with the benchmark.
- Not bothering with `SkipLocalsInit` as 64 bytes only takes 3ns to `stackalloc`, that and I don't know how it handles collection expressions 😅 
- This could be used with `ValueListBuilder.Pop` to create a stack based `Stack<Doc>` type.
  - I initially used `ValueListBuilder` and `ValueStringBuilder` for `DocFitter.Fits` but settled on the collection reuse solution.
  - The only advantage of `ValueListBuilder` using there would be instant index lookups instead of using the O(n) `ElementAt`.
  - We can reivist this if you think it would be beneficial

- I'd like to add this to `PrivatePrintLeadingTrivia` but it requires `RemoveAt` and `Insert`, `ValueListBuilder` does not implement these methods. I can add them, but I'd like to know why microsoft never added them in the first place. I hope its because they don't need it, not because its somehow unsafe.

![TokenPrivatePrint_Count](https://github.com/user-attachments/assets/ca798476-3d60-41c6-9ddd-245b00ce4e78)
![TokenPrintSyntax_Count](https://github.com/user-attachments/assets/4b427d81-c940-499d-8b72-b90fed3624fe)
Iirc `PrintTrailingSyntax` never needed a List<Doc>.



### Before
| Method                | Mean     | Error   | StdDev  | Gen0      | Gen1      | Gen2      | Allocated |
|---------------------- |---------:|--------:|--------:|----------:|----------:|----------:|----------:|
| Default_CodeFormatter | 203.2 ms | 4.06 ms | 9.33 ms | 7000.0000 | 3000.0000 | 1000.0000 |  66.17 MB |

### After
| Method                | Mean     | Error   | StdDev  | Gen0      | Gen1      | Allocated |
|---------------------- |---------:|--------:|--------:|----------:|----------:|----------:|
| Default_CodeFormatter | 169.5 ms | 3.28 ms | 4.80 ms | 5000.0000 | 2000.0000 |  58.08 MB |

Unrelated note, but when I try and run csharpier get the following message.

![image](https://github.com/user-attachments/assets/e12ce563-8f89-451e-ba5f-4d49ddb59ac6)