@belav 2025-03-01
> Unrelated note, but when I try and run csharpier get the following message.

I do have code in csharpier to automatically ignore msbuild `0.0.1`. Those CliTests were creating a csproj with msbuild `99` so if that file was still in a bin then csharpier would fail the msbuild check. The changes I merged to move the CliTest files outside of CSharpier's file structure get rid of the issue.