@belav 2025-03-01
> I'd like to add this to PrivatePrintLeadingTrivia but it requires RemoveAt and Insert, ValueListBuilder does not implement these methods. I can add them, but I'd like to know why microsoft never added them in the first place. I hope its because they don't need it, not because its somehow unsafe.

I've never loved how the code ended up handling comments and directives and some of that code is making use of Insert. At one point I was trying to simplify it but that was an excercise in frustration.

Looking at RemoveAt that seems easier to possibly work around. That logic could maybe move into the loop itself. That may be easier than implementing RemoveAt.