@TimothyMakkison 2025-03-01
> I've never loved how the code ended up handling comments and directives and some of that code is making use of Insert. At one point I was trying to simplify it but that was an excercise in frustration.
> 
> Looking at RemoveAt that seems easier to possibly work around. That logic could maybe move into the loop itself. That may be easier than implementing RemoveAt.

See #1518, `RemoveAt` is only used on the final item so I can use the MS approved `Pop` method