@TimothyMakkison 2025-03-01
> Would it make sense to use this across all of csharpier? I am pretty sure there are a ton of Lists, but this is probably the hottest path for them.

Yeah, I was going to use this for most types in `SyntaxPrinter` and `SyntaxNodePrinters` in a future PR. I wanted to keep this PR as small as possible.

> Should the constructors in Token instead guess at the capacity needed?


Do you mean something like this? 

```C#
var docs = new ValueListBuilder<Doc>(leadingTrivia.Count > 4
            ? [null, null, null, null, null,null,null,null] 
            : [null, null, null, null]);
```

>  I did see that the capacity constructor called Grow, which I assume has a bit of overhead compared to passing a Span to the constructor.

It has to call `ArrayPool` iirc this is slower than stack allocating an array.

> Most of this file is the type of code I don't really work with so I don't know how much help

Ah sorry 😓 I'll ask in the dotnets discussion page, hopefully its not a stupid question!

> I did see that the microsoft version uses scoped. Maybe including that along with params would solve the issue?

`AppendMultiChar` is indentical to the MS version they both use `scoped`. My concern is that the equation used to calculate additonal capacity appears to be incorrect in a non breaking way.