@belav 2025-03-07
> AppendMultiChar is indentical to the MS version they both use scoped. My concern is that the equation used to calculate additonal capacity appears to be incorrect/sub-optimal in a non breaking way.

Is that enough to hold off on merging this?

> Ah sorry 😓 I'll ask in the dotnets discussion page, hopefully its not a stupid question!

I am enjoying getting exposed to this type of code, but I don't know how much of it I'll retain.