@TimothyMakkison 2025-03-07
> Is that enough to hold off on merging this?

Nope, I'm not even sure if its a bug, I suspect its a slight oversight and can rarely cause a very small performance hit in rare circumstances (assuming I understand it correctly). It is nothing to worry about.

> I am enjoying getting exposed to this type of code, but I don't know how much of it I'll retain.

Glad I could help 🥰 I wouldn't worry about learning this, it's fairly niche and rarely relevant for day to day .NET work.