@belav 2025-03-07
> it's fairly niche and rarely relevant for day to day .NET work.

For sure, I'm approaching two decades working with dotnet and have had no real need to learn it yet. EF/sql is 99% of our performance issues at work.